Major fix - img_scale in config is `w*h`!!!
Back then, I found out it should be h*w, which was odd but that's what I
used. I then had problems with pipelines but I fixed those. So even
though the img_scale was incorrect, train and validation pipelines did
fix it and the models were trained properly

But the test pipeline was incorrect and deploying the model raised an
error, so this commit fixes the deployment. Thank god models were
trained properly and don't need retraining